\docType{class}
\name{BaseCorrelation-class}
\alias{BaseCorrelation-class}
\title{BaseCorrelation Class.}
\description{
A BaseCorrelation class is sude to track the results of the
baseCorrelation function.
}
\section{Slots}{
  \describe{ \item{\code{base}:}{Object of class
  \code{"matrix"}.} \item{\code{DNAPattern}:}{Object of
  class \code{"character"}} \item{\code{interp}:}{Object of
  class \code{"list"}} }
}

