\docType{class}
\name{KineticWavelets-class}
\alias{KineticWavelets-class}
\title{KineticWavelets class.}
\description{
The class is used to track a CmpH5 file and a reference
fasta to be used for further analysis in the
KineticWavelets package.
}
\section{Slots}{
  \describe{ \item{\code{h5}:}{ Object of class
  \code{"PacBioCmpH5"}.} \item{\code{reff}:}{ Object of
  class \code{"list"}.} }
}

