\name{KineticWavelets}
\alias{KineticWavelets}
\title{Constructor for KineticWavelets class}
\usage{
KineticWavelets(h5, reff)
}
\arguments{
  \item{h5}{PacBio h5 Kinetics data file}

  \item{reff}{Reference fasta file}
}
\value{
new KineticWavelets \code{linkS4class{KineticWavelets}}
}
\description{
Constructor for KineticWavelets class
}
\author{
Sterling Sawaya \email{sterlingsawaya@gmail.com}
}

