\name{plotBaseCorrelation}
\alias{plotBaseCorrelation}
\title{plotBaseCorrelation}
\usage{
plotBaseCorrelation(BaseCorrelation, ...)
}
\arguments{
  \item{BaseCorrelation}{\code{linkS4class{BaseCorrelation}}
  object}
}
\description{
Reads a BaseCorrelation object and creates a plot showing
correlation to each base pair
}

