\name{plotPatternCorrelation}
\alias{plotPatternCorrelation}
\title{plotPatternCorrelation}
\usage{
plotPatternCorrelation(BaseCorrelation, ...)
}
\arguments{
  \item{BaseCorrelation}{\code{linkS4class{BaseCorrelation}}
  object}
}
\description{
Reads a BaseCorrelation object and creates a plot showing
correlation to a DNA pattern, will return an error if
baseCorrelation was called without a DNAPattern
}

