\name{plotSmoothAverage}
\alias{plotDetailAverage}
\alias{plotSmoothAverage}
\title{plotSmoothAverage}
\usage{
plotSmoothAverage(Wave128, ...)

plotDetailAverage(Wave128, ...)
}
\arguments{
  \item{Wave128}{\code{linkS4class{Wave128}} object}

  \item{Wave128}{\code{linkS4class{Wave128}} object}
}
\description{
Reads a Wave128 object and creates a plot using the smooth
values.

Reads a Wave128 object a creates a plot using the detail
values
}

