\docType{class}
\name{Wave128-class}
\alias{Wave128-class}
\title{Wave128 Class.}
\description{
A Wave128 class is used to track the results of
wave128Window function run Object is to be passed to the
plotting functions.
}
\section{Slots}{
  \describe{ \item{\code{detailWave}:}{Object of class
  \code{"list"}.} \item{\code{smoothWave}:}{Object of class
  \code{"list"}.} \item{\code{totalElements}:}{Object of
  class \code{"numeric"}.}
  \item{\code{meanElementSize}:}{Object of class
  \code{"numeric"}.} \item{\code{alignments}:}{Object of
  class \code{"list"}.} \item{\code{DNAPattern}:}{Object of
  class \code{"character"}.}
  \item{\code{shiftWindow}:}{Object of class
  \code{"numeric"}.} \item{\code{shrink}:}{Object of class
  \code{"numeric"}.} }
}

