\name{plotPatternCorrelation}
\alias{plotPatternCorrelation}
\title{plotPatternCorrelation}
\usage{
plotPatternCorrelation(BaseCorrelation, ...)
}
\arguments{
  \item{BaseCorrelation}{\code{linkS4class{BaseCorrelation}}
  object}
}
\description{
Reads a BaseCorrelation object and creates a plot showing
correlation of kinetics with a DNA pattern.  Returns an error if
baseCorrelation was called without a DNAPattern.
}

